DO NOT MERGE
Security: tf_daemon: Update to 1.06e1

The user-space SMC Daemon is takes care in particular of the effective
storage of the SMC secure storage within the device file system. The files
in which the SMC persistent data are stored are defined in the
Normal-World Configuration File.

All the security properties such as confidentiality, integrity, atomicity,
anti-rollback and bounding to the device are managed at the SMC PA level.

Change-Id: I2b045ae77479ad90c10fc98672f7b2c2f49b68ca
Signed-off-by: Bryan Buckley <bryan.buckley@ti.com>